create seperate context for reset password
