+++ src/apps/frontend/app.component.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { Toaster } from 'react-hot-toast';
 import { BrowserRouter as Router } from 'react-router-dom';
 
-import { AccountProvider, AuthProvider } from './contexts';
+import { AccountProvider, AuthProvider, ResetPasswordProvider } from './contexts';
 import { Config } from './helpers';
 import { AppRoutes } from './routes';
 import InspectLet from './vendor/inspectlet';
@@ -18,12 +18,14 @@ export default function App(): React.ReactElement {
 
   return (
     <AuthProvider>
-      <AccountProvider>
-        <Toaster />
-        <Router>
-          <AppRoutes />
-        </Router>
-      </AccountProvider>
+      <ResetPasswordProvider>
+        <AccountProvider>
+          <Toaster />
+          <Router>
+            <AppRoutes />
+          </Router>
+        </AccountProvider>
+      </ResetPasswordProvider>
     </AuthProvider>
   );
 }
