+++ src/apps/frontend/contexts/index.ts
@@ -1,7 +1,9 @@
 import { AccountProvider, useAccountContext } from './account.provider';
 import { AuthProvider, useAuthContext } from './auth.provider';
+import { ResetPasswordProvider, useResetPasswordContext } from './reset-password.provider';
 
 export {
   AuthProvider, useAuthContext,
   AccountProvider, useAccountContext,
+  ResetPasswordProvider, useResetPasswordContext,
 };
