+++ src/apps/frontend/services/auth.service.ts
@@ -22,19 +22,4 @@ export default class AuthService extends APIService {
     username,
     password,
   });
-
-  sendForgotPasswordEmail = async (
-    username: string,
-  ): Promise<ApiResponse<void>> => this.apiClient.post('/password-reset-token', {
-    username,
-  });
-
-  resetPassword = async (
-    accountId: string,
-    newPassword: string,
-    token: string,
-  ): Promise<ApiResponse<void>> => this.apiClient.patch(`/accounts/${accountId}/reset-password`, {
-    newPassword,
-    token,
-  });
 }
