+++ src/apps/frontend/services/index.ts
@@ -1,9 +1,11 @@
 import AccountService from './account.service';
 import APIService from './api.service';
 import AuthService from './auth.service';
+import ResetPasswordService from './reset-password.service';
 
 export {
   AccountService,
   AuthService,
   APIService,
+  ResetPasswordService,
 };
